[Resource] unify database resource
