+++ config/template.yaml
@@ -25,6 +25,7 @@ plugins:
     memory:
       type: pipeline.plugins.resources.memory:MemoryResource
       database:
+        # DatabaseResource implementation
         type: pipeline.plugins.resources.sqlite_storage:SQLiteStorageResource
         path: ./agent.db
       filesystem:
